refactor(currency): change class name to calulator -> calculator
