+++ src/main/java/com/exchange/calculator/CalculatorApplication.java
@@ -1,13 +1,13 @@
-package com.exchange.calulator;
+package com.exchange.calculator;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class CalulatorApplication {
+public class CalculatorApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(CalulatorApplication.class, args);
+        SpringApplication.run(CalculatorApplication.class, args);
     }
 
 }
