+++ src/main/java/com/exchange/calculator/currency/infra/CurrencyRateClient.java
@@ -1,0 +1,7 @@
+package com.exchange.calculator.currency.infra;
+
+import com.exchange.calculator.currency.dto.CurrencyResponse;
+
+public interface CurrencyRateClient {
+    CurrencyResponse retrieveCurrency();
+}
