+++ src/main/java/com/exchange/calculator/currency/ui/CurrencyRestController.java
@@ -1,6 +1,8 @@
-package com.exchange.calulator.currency.ui;
+package com.exchange.calculator.currency.ui;
 
-import com.exchange.calulator.currency.dto.CurrencyResponse;
+import com.exchange.calculator.currency.application.CurrencyApplication;
+import com.exchange.calculator.currency.dto.CurrencyResponse;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -8,10 +10,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/currency")
+@RequiredArgsConstructor
 public class CurrencyRestController {
+
+    private final CurrencyApplication currencyApplication;
 
     @GetMapping
     public CurrencyResponse getCurrency(@RequestParam(value = "source", required = false) String source) {
-        return null;
+        return currencyApplication.getCurrencyData();
     }
 }
