--- src/main/java/com/exchange/calulator/currency/infra/CurrencyRateClient.java
@@ -1,8 +1,0 @@
-package com.exchange.calulator.currency.infra;
-
-import com.exchange.calulator.currency.dto.CurrencyResponse;
-import com.fasterxml.jackson.core.JsonProcessingException;
-
-public interface CurrencyRateClient {
-    CurrencyResponse retrieveCurrency() throws JsonProcessingException;
-}
